pyproject.toml
[tool.poetry]
name = "skillpacks"
version = "0.1.0"
description = "Pluggable skills for AI agents"
authors = ["Patrick Barker <patrickbarkerco@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.1"
sqlalchemy = "^2.0.27"
agentdesk = "^0.2.7"
mllm = "^0.1.3"


[tool.poetry.group.openai.dependencies]
openai = "^1.12.0"


[tool.poetry.group.qwen.dependencies]
transformers = "^4.38.1"
frequency-ai = "^0.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
